+++ buildSrc/src/main/java/com/google/firebase/gradle/plugins/BaseFirebaseLibraryPlugin.kt
@@ -189,7 +189,7 @@ abstract class BaseFirebaseLibraryPlugin : Plugin<Project> {
   // TODO(b/277607560): Remove when Gradle's MavenPublishPlugin adds functionality for aar types
   private fun addTypeWithAARSupport(dependency: Element, androidLibraries: List<String>) {
     dependency.findOrCreate("type").apply {
-      textContent = if (androidLibraries.contains(dependency.toArtifactString())) "aar" else "jar"
+      textContent = if (androidLibraries.contains(dependency.toMavenName())) "aar" else "jar"
     }
   }
 }
@@ -200,7 +200,7 @@ abstract class BaseFirebaseLibraryPlugin : Plugin<Project> {
  * This is collected via the [runtimeClasspath][FirebaseLibraryExtension.getRuntimeClasspath], and
  * includes project level dependencies as well as external dependencies.
  *
- * The dependencies are mapped to their [artifactName].
+ * The dependencies are mapped to their [mavenName][toMavenName].
  *
  * @see resolveProjectLevelDependencies
  * @see resolveExternalAndroidLibraries
@@ -208,9 +208,7 @@ abstract class BaseFirebaseLibraryPlugin : Plugin<Project> {
 // TODO(b/277607560): Remove when Gradle's MavenPublishPlugin adds functionality for aar types
 fun FirebaseLibraryExtension.resolveAndroidDependencies() =
   resolveExternalAndroidLibraries() +
-    resolveProjectLevelDependencies()
-      .filter { it.type == LibraryType.ANDROID }
-      .map { it.artifactName }
+    resolveProjectLevelDependencies().filter { it.type == LibraryType.ANDROID }.map { it.mavenName }
 
 /**
  * A list of project level dependencies.
@@ -239,9 +237,9 @@ fun FirebaseLibraryExtension.resolveProjectLevelDependencies() =
  * This is collected via the [runtimeClasspath][FirebaseLibraryExtension.getRuntimeClasspath], using
  * an [ArtifactView][org.gradle.api.artifacts.ArtifactView] that filters for `aar` artifactType.
  *
- * Artifacts are mapped to their respective display name:
+ * Artifacts are mapped to their respective maven name:
  * ```
- * groupId:artifactId:version
+ * groupId:artifactId
  * ```
  */
 // TODO(b/277607560): Remove when Gradle's MavenPublishPlugin adds functionality for aar types
@@ -251,7 +249,7 @@ fun FirebaseLibraryExtension.resolveExternalAndroidLibraries() =
     .incoming
     .artifactView { attributes { attribute("artifactType", "aar") } }
     .artifacts
-    .map { it.variant.displayName.substringBefore(" ") }
+    .map { it.variant.displayName.substringBefore(" ").substringBeforeLast(":") }
 
 /**
  * The name provided to this artifact when published.