+++ appcheck/firebase-appcheck/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 * [feature] Added `getLimtedUseAppCheckToken()` for obtaining limited-use tokens
-  for protecting non-Firebase backends.
+  for protecting non-Firebase backends. [#4876](//github.com/firebase/firebase-android-sdk/issues/4876){: .external})
+* [changed] Internal changes to ensure alignment with other SDK releases.
 
 # 16.1.2
 * [unchanged] Updated to keep [app_check] SDK versions aligned.